move the `encode()` methods to the `Request` and `Response` classes
the actual encoding is still done in `BasicStructure`, but within
`encode_into_pdu()`.
